Move file banner logic to the new system
